refactor(server): split routes.rs by route family (TD-033)
Behavior-preserving decomposition of parish-server/src/routes.rs (~3300 LOC)
into submodules by route family:

- routes/world.rs     — query endpoints (snapshot, map, npcs, theme, debug, bug report)
- routes/input.rs     — player input and game-loop adapters
- routes/reactions.rs — NPC reaction recording and emission
- routes/saves.rs     — save-file and branch lifecycle
- routes/admin.rs     — admin guard, branch-name validation, addressed_to validation
- routes/demo.rs      — demo/screenshot stubs (desktop-only, returns 501)
- routes/mods.rs      — mod listing and switching
- routes/session_token.rs — WS session-token issuance

routes/mod.rs is the hub: re-exports all public items so lib.rs route
registration and all external use paths (parish_server::routes::*) are
unchanged. Tests moved alongside with full imports restored.

Part of #1200 (TD-033). Does NOT close the epic.
