+++ parish/crates/parish-server/src/routes/mod.rs
@@ -1,0 +1,1601 @@
+//! HTTP route handlers for the Parish web server.
+//!
+//! Each route maps to a Tauri command, calling the shared handlers in
+//! [`parish_core::ipc`] and returning JSON responses.
+//!
+//! Handlers are split into submodules by route family:
+//! - [`world`]         — query endpoints (snapshot, map, npcs, theme, debug, bug report)
+//! - [`input`]         — player input and game-loop adapters
+//! - [`reactions`]     — NPC reaction recording and emission
+//! - [`saves`]         — save-file and branch lifecycle
+//! - [`admin`]         — admin guard, branch-name validation, addressed_to validation
+//! - [`demo`]          — demo/screenshot stubs (desktop-only, returns 501)
+//! - [`mods`]          — mod listing and switching
+//! - [`session_token`] — WS session-token issuance
+
+pub mod admin;
+pub mod demo;
+pub mod input;
+pub mod mods;
+pub mod reactions;
+pub mod saves;
+pub mod session_token;
+pub mod world;
+
+// ── Re-exports — keep lib.rs route registration unchanged ────────────────────
+
+// world
+pub use world::{
+    build_full_debug_snapshot, get_app_icon, get_available_providers, get_debug_snapshot,
+    get_favicon, get_health, get_map, get_npcs_here, get_setup_snapshot, get_theme, get_ui_config,
+    get_world_snapshot, redact_call_log, serve_mod_icon, submit_bug_report,
+};
+
+// input
+pub use input::{
+    SubmitInputRequest, emit_world_update, handle_game_input, handle_system_command,
+    rebuild_inference_inner, spawn_loading_animation, submit_input, tick_inactivity,
+    touch_player_activity,
+};
+
+// reactions
+pub use reactions::{emit_npc_reactions, react_to_message};
+
+// saves
+pub use saves::{
+    CreateBranchRequest, LoadBranchRequest, create_branch, discover_save_files,
+    do_branch_log_inner, do_fork_branch_inner, do_list_branches_inner, do_new_game_inner,
+    do_save_game_inner, get_save_state, load_branch, load_branch_snapshot, new_game, new_save_file,
+    restore_snapshot_and_emit, save_game, validate_and_acquire_lock,
+};
+
+// admin
+pub use admin::{
+    admin_emails, check_admin, check_admin_against, check_admin_no_config, is_admin_command,
+    parse_admin_emails, validate_addressed_to, validate_branch_name,
+};
+
+// demo
+pub use demo::{
+    get_demo_config, get_demo_context, get_latest_screenshot, get_llm_player_action,
+    save_screenshot, take_screenshot,
+};
+
+// mods
+pub use mods::{ModEntry, SwitchModBody, collect_base_mods, list_mods, mods_root_path, switch_mod};
+
+// session_token
+pub use session_token::{SessionInitResponse, session_init};
+
+// ── Tests ────────────────────────────────────────────────────────────────────
+
+#[cfg(test)]
+pub mod tests {
+    // Re-export submodule items used by tests so they resolve via `super::`.
+    use super::*;
+    use std::collections::VecDeque;
+    use std::sync::{Arc, Mutex as StdMutex};
+    use std::time::{Duration, Instant};
+
+    use axum::Json;
+    use axum::response::IntoResponse;
+    use parish_core::event_bus::EventBus as EventBusTrait;
+
+    use parish_core::game_loop::is_snippet_injection_char;
+    use parish_core::inference::{InferenceQueue, InferenceRequest, InferenceResponse};
+    use parish_core::ipc::capitalize_first;
+    use parish_core::ipc::{NpcReactionPayload, TextLogPayload};
+    use parish_core::npc::Npc;
+    use parish_core::npc::manager::NpcManager;
+    use parish_core::world::transport::TransportConfig;
+    use parish_core::world::{DEFAULT_START_LOCATION, LocationId, WorldState};
+    use tower::ServiceExt;
+
+    #[cfg(test)]
+    use parish_core::ipc::ConversationLine;
+    #[cfg(test)]
+    use parish_core::npc::NpcId;
+    #[cfg(test)]
+    use tokio::sync::mpsc;
+
+    #[test]
+    fn submit_input_request_deserialization() {
+        let json = r#"{"text": "go to church"}"#;
+        let req: SubmitInputRequest = serde_json::from_str(json).unwrap();
+        assert_eq!(req.text, "go to church");
+        assert!(req.addressed_to.is_empty());
+    }
+
+    #[test]
+    fn submit_input_request_with_addressed_to() {
+        let json = r#"{"text": "hello", "addressedTo": ["Padraig", "Maire"]}"#;
+        let req: SubmitInputRequest = serde_json::from_str(json).unwrap();
+        assert_eq!(req.text, "hello");
+        assert_eq!(req.addressed_to, vec!["Padraig", "Maire"]);
+    }
+
+    #[test]
+    fn parse_admin_emails_basic_list() {
+        let set = parse_admin_emails("alice@example.com,bob@example.com");
+        assert!(set.contains("alice@example.com"));
+        assert!(set.contains("bob@example.com"));
+        assert_eq!(set.len(), 2);
+    }
+
+    #[test]
+    fn parse_admin_emails_trims_and_drops_empties() {
+        let set = parse_admin_emails(" alice@example.com , , bob@example.com ,");
+        assert!(set.contains("alice@example.com"));
+        assert!(set.contains("bob@example.com"));
+        assert_eq!(
+            set.len(),
+            2,
+            "empty entries and surrounding spaces must be dropped"
+        );
+    }
+
+    #[test]
+    fn parse_admin_emails_empty_string_returns_empty_set() {
+        let set = parse_admin_emails("");
+        assert!(set.is_empty());
+    }
+
+    /// Helper to build a minimal AppState from the real game data.
+    pub fn test_app_state() -> Arc<crate::state::AppState> {
+        let data_dir =
+            std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../../mods/rundale");
+        let world =
+            WorldState::from_parish_file(&data_dir.join("world.json"), DEFAULT_START_LOCATION)
+                .unwrap();
+        let npc_manager = NpcManager::new();
+        let transport = TransportConfig::default();
+        let ui_config = crate::state::UiConfigSnapshot {
+            hints_label: "test".to_string(),
+            default_accent: "#000".to_string(),
+            splash_text: String::new(),
+            active_tile_source: String::new(),
+            tile_sources: Vec::new(),
+            auto_pause_timeout_seconds: 300,
+            app_icon_url: None,
+            favicon_url: None,
+            map_overlay: None,
+            base_mod_required: false,
+        };
+        let theme_palette = parish_core::game_mod::default_theme_palette();
+        let saves_dir = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../../saves");
+        let session_store: std::sync::Arc<dyn parish_core::session_store::SessionStore> =
+            std::sync::Arc::new(crate::session_store_impl::DbSessionStore::new(
+                saves_dir.clone(),
+            ));
+        crate::state::build_app_state(
+            "test-session".to_string(),
+            world,
+            npc_manager,
+            None,
+            crate::state::GameConfig {
+                provider_name: String::new(),
+                base_url: String::new(),
+                api_key: None,
+                model_name: String::new(),
+                cloud_provider_name: None,
+                cloud_model_name: None,
+                cloud_api_key: None,
+                cloud_base_url: None,
+                improv_enabled: false,
+                max_follow_up_turns: 2,
+                idle_banter_after_secs: 25,
+                auto_pause_after_secs: 60,
+                category_provider: Default::default(),
+                category_model: Default::default(),
+                category_api_key: Default::default(),
+                category_base_url: Default::default(),
+                flags: parish_core::config::FeatureFlags::default(),
+                category_rate_limit: Default::default(),
+                active_tile_source: String::new(),
+                tile_sources: Vec::new(),
+                reveal_unexplored_locations: false,
+                auto_setup_model: None,
+            },
+            None,
+            transport,
+            ui_config,
+            theme_palette,
+            saves_dir,
+            data_dir.clone(),
+            None,
+            data_dir.join("parish-flags.json"),
+            parish_core::config::InferenceConfig::default(),
+            session_store,
+            parish_core::inference::file_log::InferenceFileLog::disabled(),
+            parish_core::chat_transcript::ChatTranscriptLog::disabled(),
+        )
+    }
+
+    /// #1164 AC1: `GET /api/world-snapshot` (the endpoint the reconnect resync
+    /// re-fetches) must report `turn_in_flight` from the authoritative
+    /// conversation state so the web client can re-assert `streamingActive`
+    /// instead of clearing it mid-turn.
+    #[tokio::test]
+    async fn world_snapshot_reports_turn_in_flight_from_conversation_state() {
+        let state = test_app_state();
+
+        // Idle: no turn in flight.
+        let Json(idle) =
+            super::get_world_snapshot(axum::extract::Extension(Arc::clone(&state))).await;
+        assert!(
+            !idle.turn_in_flight,
+            "expected turn_in_flight=false when idle"
+        );
+
+        // Simulate an NPC turn being processed.
+        state.conversation.lock().await.conversation_in_progress = true;
+        let Json(busy) =
+            super::get_world_snapshot(axum::extract::Extension(Arc::clone(&state))).await;
+        assert!(
+            busy.turn_in_flight,
+            "expected turn_in_flight=true while a conversation turn is in flight"
+        );
+
+        // Turn finishes: signal clears again.
+        state.conversation.lock().await.conversation_in_progress = false;
+        let Json(done) =
+            super::get_world_snapshot(axum::extract::Extension(Arc::clone(&state))).await;
+        assert!(
+            !done.turn_in_flight,
+            "expected turn_in_flight=false after the turn completes"
+        );
+    }
+
+    async fn add_introduced_npc(
+        state: &Arc<crate::state::AppState>,
+        id: u32,
+        name: &str,
+        occupation: &str,
+    ) {
+        let player_location = {
+            let world = state.world.lock().await;
+            world.player_location
+        };
+
+        let mut npc = Npc::new_test_npc();
+        npc.id = NpcId(id);
+        npc.name = name.to_string();
+        npc.occupation = occupation.to_string();
+        npc.brief_description = format!("a {}", occupation.to_lowercase());
+        npc.location = player_location;
+
+        let mut npc_manager = state.npc_manager.lock().await;
+        npc_manager.add_npc(npc);
+        npc_manager.mark_introduced(NpcId(id));
+    }
+
+    async fn install_scripted_inference_queue(
+        state: &Arc<crate::state::AppState>,
+        responses: Vec<&str>,
+    ) -> (Arc<StdMutex<Vec<String>>>, tokio::task::JoinHandle<()>) {
+        let (tx, mut rx) = mpsc::channel::<InferenceRequest>(8);
+        let (bg_tx, _bg_rx) = mpsc::channel::<InferenceRequest>(8);
+        let (batch_tx, _batch_rx) = mpsc::channel::<InferenceRequest>(8);
+        let prompts = Arc::new(StdMutex::new(Vec::new()));
+        let prompt_log = Arc::clone(&prompts);
+        let mut scripted: VecDeque<String> = responses.into_iter().map(str::to_string).collect();
+
+        let handle = tokio::spawn(async move {
+            while let Some(request) = rx.recv().await {
+                prompt_log.lock().unwrap().push(request.prompt.clone());
+
+                let text = scripted.pop_front().unwrap_or_else(|| {
+                    r#"{"dialogue":"Aye.","action":"speaks","mood":"content"}"#.to_string()
+                });
+
+                let _ = request.response_tx.send(InferenceResponse {
+                    id: request.id,
+                    text,
+                    error: None,
+                });
+            }
+        });
+
+        *state.inference_queue.lock().await = Some(InferenceQueue::new(tx, bg_tx, batch_tx));
+        (prompts, handle)
+    }
+
+    fn drain_text_logs(stream: &mut parish_core::event_bus::EventStream) -> Vec<TextLogPayload> {
+        let mut logs = Vec::new();
+        loop {
+            match stream.try_recv() {
+                Some(event) if event.event == "text-log" => {
+                    logs.push(serde_json::from_value(event.payload).unwrap());
+                }
+                Some(_) => {}
+                None => break,
+            }
+        }
+        logs
+    }
+
+    /// Verifies that handle_movement resolves and applies movement atomically
+    /// (clock advance + player_location update within a single lock scope).
+    #[tokio::test]
+    async fn handle_movement_updates_location_and_clock() {
+        let state = test_app_state();
+
+        let (start_loc, start_time) = {
+            let world = state.world.lock().await;
+            (world.player_location, world.clock.now())
+        };
+
+        // Move to the crossroads (a neighbor of Kilteevan Village, id 15)
+        input::handle_movement("crossroads", &state).await;
+
+        let world = state.world.lock().await;
+        assert_ne!(
+            world.player_location, start_loc,
+            "player_location should change after movement"
+        );
+        // Clock should have advanced (travel takes > 0 minutes)
+        assert!(
+            world.clock.now() > start_time,
+            "clock should advance during travel"
+        );
+    }
+
+    /// Verifies that moving to an unknown location does not change world state.
+    #[tokio::test]
+    async fn handle_movement_unknown_destination_preserves_state() {
+        let state = test_app_state();
+
+        let (start_loc, start_time) = {
+            let mut world = state.world.lock().await;
+            world.clock.pause();
+            (world.player_location, world.clock.now())
+        };
+
+        input::handle_movement("nonexistent-place-xyz", &state).await;
+
+        let world = state.world.lock().await;
+        assert_eq!(
+            world.player_location, start_loc,
+            "player_location should not change for unknown destination"
+        );
+        assert_eq!(
+            world.clock.now(),
+            start_time,
+            "clock should not advance for unknown destination"
+        );
+    }
+
+    #[test]
+    fn text_log_generates_unique_ids() {
+        let a = parish_core::ipc::text_log("system", "hello");
+        let b = parish_core::ipc::text_log("system", "world");
+        assert_ne!(a.id, b.id);
+        assert!(a.id.starts_with("msg-"));
+    }
+
+    #[test]
+    fn react_request_deserialization() {
+        let json = r#"{"npcName": "Padraig", "messageSnippet": "Hello", "emoji": "😊"}"#;
+        let req: parish_core::ipc::ReactRequest = serde_json::from_str(json).unwrap();
+        assert_eq!(req.npc_name, "Padraig");
+        assert_eq!(req.emoji, "😊");
+    }
+
+    /// Verifies that get_save_state returns None fields on fresh AppState.
+    #[tokio::test]
+    async fn get_save_state_initial_is_empty() {
+        let state = test_app_state();
+        let result = get_save_state(axum::extract::Extension(state)).await;
+        let save_state = result.0;
+        assert!(save_state.filename.is_none());
+        assert!(save_state.branch_id.is_none());
+        assert!(save_state.branch_name.is_none());
+    }
+
+    /// Verifies that discover_save_files returns an empty list for a missing saves dir.
+    #[tokio::test]
+    async fn discover_save_files_empty_dir() {
+        let state = test_app_state();
+        // saves_dir points to ../../saves which may or may not exist — either way should not panic
+        let result = discover_save_files(axum::extract::Extension(state)).await;
+        assert!(result.is_ok());
+    }
+
+    /// Verifies request body deserialization for load_branch.
+    #[test]
+    fn load_branch_request_deserialization() {
+        let json = r#"{"filePath": "/saves/parish_001.db", "branchId": 1}"#;
+        let req: LoadBranchRequest = serde_json::from_str(json).unwrap();
+        assert_eq!(req.file_path, "/saves/parish_001.db");
+        assert_eq!(req.branch_id, 1);
+    }
+
+    /// Verifies request body deserialization for create_branch.
+    #[test]
+    fn create_branch_request_deserialization() {
+        let json = r#"{"name": "alternate", "parentBranchId": 1}"#;
+        let req: CreateBranchRequest = serde_json::from_str(json).unwrap();
+        assert_eq!(req.name, "alternate");
+        assert_eq!(req.parent_branch_id, 1);
+    }
+
+    #[tokio::test]
+    async fn handle_npc_conversation_preserves_order_and_follow_up_context() {
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+        add_introduced_npc(&state, 2, "Padraig Darcy", "Farmer").await;
+
+        {
+            let mut config = state.config.lock().await;
+            config.model_name = "test-model".to_string();
+            config.max_follow_up_turns = 0;
+        }
+
+        // Subscribe BEFORE the dispatch so we can count stream-end events.
+        let mut rx = state.event_bus.subscribe(&[]);
+
+        let (prompts, worker) = install_scripted_inference_queue(
+            &state,
+            vec![
+                r#"{"dialogue":"I heard the fair will be lively.","action":"speaks","mood":"curious"}"#,
+                r#"{"dialogue":"If it is, Siobhan, I'll bring the cart.","action":"speaks","mood":"content"}"#,
+            ],
+        )
+        .await;
+
+        input::handle_npc_conversation(
+            "What news is there?".to_string(),
+            vec!["Siobhan Murphy".to_string(), "Padraig Darcy".to_string()],
+            &state,
+        )
+        .await;
+
+        let transcript = {
+            let conversation = state.conversation.lock().await;
+            conversation.transcript.iter().cloned().collect::<Vec<_>>()
+        };
+        assert_eq!(
+            transcript,
+            vec![
+                ConversationLine {
+                    speaker: "You".to_string(),
+                    text: "What news is there?".to_string(),
+                },
+                ConversationLine {
+                    speaker: "Siobhan Murphy".to_string(),
+                    text: "I heard the fair will be lively.".to_string(),
+                },
+                ConversationLine {
+                    speaker: "Padraig Darcy".to_string(),
+                    text: "If it is, Siobhan, I'll bring the cart.".to_string(),
+                },
+            ]
+        );
+
+        let prompts = prompts.lock().unwrap().clone();
+        assert_eq!(prompts.len(), 2);
+        // First prompt: the player's current input is excluded from the "Recent
+        // conversation" section (it's shown separately as the triggering line via
+        // build_named_action_line), so no transcript header appears.
+        assert!(prompts[0].contains("The newcomer says: \"What news is there?\""));
+        // Second prompt: includes Siobhan's prior response in transcript context.
+        assert!(prompts[1].contains("Recent conversation here:"));
+        assert!(prompts[1].contains("- Siobhan Murphy: I heard the fair will be lively."));
+
+        // Regression guard: stream-end must fire EXACTLY ONCE for the whole
+        // turn (addressed + follow-up), so the input field stays disabled
+        // through every NPC's response. PR #222 emitted one per turn, which
+        // let the input flicker open between NPCs and contradicted the
+        // explicit user spec.
+        let mut stream_end_count = 0;
+        loop {
+            match rx.try_recv() {
+                Some(event) if event.event == "stream-end" => stream_end_count += 1,
+                Some(_) => {}
+                None => break,
+            }
+        }
+        assert_eq!(
+            stream_end_count, 1,
+            "expected exactly one stream-end after a 2-turn dispatch, got {}",
+            stream_end_count
+        );
+
+        worker.abort();
+    }
+
+    /// #1164 AC2: every `stream-token` emitted for a player-initiated NPC
+    /// conversation turn must carry the same non-empty `message_id` as the
+    /// turn's `text-log` placeholder, so a stream that resumes after a
+    /// WebSocket reconnect can rebind to a reactable chat entry.
+    #[tokio::test]
+    async fn stream_tokens_carry_the_placeholder_message_id() {
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+
+        let mut rx = state.event_bus.subscribe(&[]);
+
+        // A streaming worker: unlike `install_scripted_inference_queue`, this
+        // pushes tokens through `token_tx` so `stream-token` events are
+        // actually emitted (the scripted helper only sends the final response).
+        let (tx, mut req_rx) = mpsc::channel::<InferenceRequest>(8);
+        let (bg_tx, _bg_rx) = mpsc::channel::<InferenceRequest>(8);
+        let (batch_tx, _batch_rx) = mpsc::channel::<InferenceRequest>(8);
+        let worker = tokio::spawn(async move {
+            while let Some(request) = req_rx.recv().await {
+                if let Some(token_tx) = &request.token_tx {
+                    let _ = token_tx
+                        .send("Aye, the fair will be grand.".to_string())
+                        .await;
+                }
+                let _ = request.response_tx.send(InferenceResponse {
+                    id: request.id,
+                    text: r#"{"dialogue":"Aye, the fair will be grand.","action":"speaks","mood":"content"}"#
+                        .to_string(),
+                    error: None,
+                });
+            }
+        });
+        *state.inference_queue.lock().await = Some(InferenceQueue::new(tx, bg_tx, batch_tx));
+
+        input::handle_npc_conversation(
+            "What news is there?".to_string(),
+            vec!["Siobhan Murphy".to_string()],
+            &state,
+        )
+        .await;
+
+        // Collect the placeholder id (from the empty streaming text-log) and the
+        // ids carried by stream-token events for that turn.
+        let mut placeholder_id: Option<String> = None;
+        let mut placeholder_turn: Option<u64> = None;
+        let mut token_message_ids: Vec<(u64, Option<String>)> = Vec::new();
+        loop {
+            match rx.try_recv() {
+                Some(event) if event.event == "text-log" => {
+                    let p = &event.payload;
+                    if p.get("content").and_then(|v| v.as_str()) == Some("")
+                        && p.get("stream_turn_id").and_then(|v| v.as_u64()).is_some()
+                    {
+                        placeholder_turn = p.get("stream_turn_id").and_then(|v| v.as_u64());
+                        placeholder_id = p.get("id").and_then(|v| v.as_str()).map(str::to_string);
+                    }
+                }
+                Some(event) if event.event == "stream-token" => {
+                    let turn = event
+                        .payload
+                        .get("turn_id")
+                        .and_then(|v| v.as_u64())
+                        .unwrap_or_default();
+                    let mid = event
+                        .payload
+                        .get("message_id")
+                        .and_then(|v| v.as_str())
+                        .map(str::to_string);
+                    token_message_ids.push((turn, mid));
+                }
+                Some(_) => {}
+                None => break,
+            }
+        }
+
+        let placeholder_id = placeholder_id.expect("expected a streaming text-log placeholder");
+        assert!(
+            !placeholder_id.is_empty(),
+            "placeholder id must be non-empty"
+        );
+        assert!(
+            !token_message_ids.is_empty(),
+            "expected at least one stream-token for the turn"
+        );
+        for (turn, mid) in &token_message_ids {
+            assert_eq!(
+                Some(*turn),
+                placeholder_turn,
+                "stream-token turn_id should match the placeholder's stream_turn_id"
+            );
+            assert_eq!(
+                mid.as_deref(),
+                Some(placeholder_id.as_str()),
+                "every stream-token must carry the placeholder's message_id (#1164)"
+            );
+        }
+
+        worker.abort();
+    }
+
+    #[tokio::test]
+    async fn handle_npc_conversation_bystander_chain_picks_related_npc() {
+        use parish_core::npc::types::{Relationship, RelationshipKind};
+
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+        add_introduced_npc(&state, 2, "Padraig Darcy", "Farmer").await;
+        add_introduced_npc(&state, 3, "Sean Brennan", "Smith").await;
+
+        // Sean has a strong friendship with Padraig — when Padraig is the last
+        // speaker, the heuristic should pick Sean for the autonomous chain
+        // turn (not Siobhan, who has already spoken and is excluded by
+        // `recently_spoken`).
+        {
+            let mut npc_manager = state.npc_manager.lock().await;
+            if let Some(sean) = npc_manager.get_mut(NpcId(3)) {
+                sean.relationships.insert(
+                    NpcId(2),
+                    Relationship {
+                        kind: RelationshipKind::Friend,
+                        strength: 0.7,
+                        history: Vec::new(),
+                    },
+                );
+            }
+        }
+
+        {
+            let mut config = state.config.lock().await;
+            config.model_name = "test-model".to_string();
+            config.max_follow_up_turns = 1;
+            // Bystander chain is gated behind `autonomous-npc-chain` (off by
+            // default); enable it so this regression guard exercises Phase 2.
+            config
+                .flags
+                .enable(parish_core::game_loop::AUTONOMOUS_NPC_CHAIN_FLAG);
+        }
+
+        let (_prompts, worker) = install_scripted_inference_queue(
+            &state,
+            vec![
+                r#"{"dialogue":"I heard the fair will be lively.","action":"speaks","mood":"curious"}"#,
+                r#"{"dialogue":"If it is, Siobhan, I'll bring the cart.","action":"speaks","mood":"content"}"#,
+                r#"{"dialogue":"I'd come too if my hand wasn't burnt at the forge.","action":"speaks","mood":"content"}"#,
+            ],
+        )
+        .await;
+
+        input::handle_npc_conversation(
+            "What news is there?".to_string(),
+            vec!["Siobhan Murphy".to_string(), "Padraig Darcy".to_string()],
+            &state,
+        )
+        .await;
+
+        let transcript = {
+            let conversation = state.conversation.lock().await;
+            conversation.transcript.iter().cloned().collect::<Vec<_>>()
+        };
+        // Expect: player → Siobhan (addressed) → Padraig (addressed) → Sean (chain).
+        assert_eq!(transcript.len(), 4, "transcript = {:?}", transcript);
+        assert_eq!(transcript[3].speaker, "Sean Brennan");
+
+        worker.abort();
+    }
+
+    /// AC1 — Default off: with no flag set and bystanders present, only the
+    /// addressed NPC replies. The chain that the always-on behaviour would
+    /// have produced must not fire.
+    #[tokio::test]
+    async fn handle_npc_conversation_chain_disabled_by_default() {
+        use parish_core::npc::types::{Relationship, RelationshipKind};
+
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+        add_introduced_npc(&state, 2, "Padraig Darcy", "Farmer").await;
+
+        // High-strength friendship — under the old always-on chain this would
+        // have lured Padraig into Phase 2. Off-by-default must suppress it.
+        {
+            let mut npc_manager = state.npc_manager.lock().await;
+            if let Some(padraig) = npc_manager.get_mut(NpcId(2)) {
+                padraig.relationships.insert(
+                    NpcId(1),
+                    Relationship {
+                        kind: RelationshipKind::Friend,
+                        strength: 0.9,
+                        history: Vec::new(),
+                    },
+                );
+            }
+        }
+
+        {
+            let mut config = state.config.lock().await;
+            config.model_name = "test-model".to_string();
+            config.max_follow_up_turns = 2;
+            // No `flags.enable(...)` — relying on default-off.
+            assert!(
+                !config
+                    .flags
+                    .is_enabled(parish_core::game_loop::AUTONOMOUS_NPC_CHAIN_FLAG),
+                "flag must default to disabled"
+            );
+        }
+
+        let (_prompts, worker) = install_scripted_inference_queue(
+            &state,
+            // Only one reply is needed when the chain is gated off.
+            vec![r#"{"dialogue":"I heard the fair will be lively.","action":"speaks","mood":"curious"}"#],
+        )
+        .await;
+
+        input::handle_npc_conversation(
+            "What news is there?".to_string(),
+            vec!["Siobhan Murphy".to_string()],
+            &state,
+        )
+        .await;
+
+        let transcript = {
+            let conversation = state.conversation.lock().await;
+            conversation.transcript.iter().cloned().collect::<Vec<_>>()
+        };
+        // Expect: player → Siobhan only. No Padraig chain turn.
+        assert_eq!(transcript.len(), 2, "transcript = {:?}", transcript);
+        assert_eq!(transcript[1].speaker, "Siobhan Murphy");
+
+        worker.abort();
+    }
+
+    /// AC3 — Explicit disable: `flags.disable(...)` behaves identically to
+    /// the never-set default. Guards against an accidental coupling where
+    /// `is_disabled` ever became the gating predicate.
+    #[tokio::test]
+    async fn handle_npc_conversation_chain_explicit_disable_matches_default() {
+        use parish_core::npc::types::{Relationship, RelationshipKind};
+
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+        add_introduced_npc(&state, 2, "Padraig Darcy", "Farmer").await;
+
+        {
+            let mut npc_manager = state.npc_manager.lock().await;
+            if let Some(padraig) = npc_manager.get_mut(NpcId(2)) {
+                padraig.relationships.insert(
+                    NpcId(1),
+                    Relationship {
+                        kind: RelationshipKind::Friend,
+                        strength: 0.9,
+                        history: Vec::new(),
+                    },
+                );
+            }
+        }
+
+        {
+            let mut config = state.config.lock().await;
+            config.model_name = "test-model".to_string();
+            config.max_follow_up_turns = 2;
+            config
+                .flags
+                .disable(parish_core::game_loop::AUTONOMOUS_NPC_CHAIN_FLAG);
+        }
+
+        let (_prompts, worker) = install_scripted_inference_queue(
+            &state,
+            vec![r#"{"dialogue":"Quiet at the fair.","action":"speaks","mood":"content"}"#],
+        )
+        .await;
+
+        input::handle_npc_conversation(
+            "Anything stirring?".to_string(),
+            vec!["Siobhan Murphy".to_string()],
+            &state,
+        )
+        .await;
+
+        let transcript = {
+            let conversation = state.conversation.lock().await;
+            conversation.transcript.iter().cloned().collect::<Vec<_>>()
+        };
+        assert_eq!(transcript.len(), 2, "transcript = {:?}", transcript);
+        assert_eq!(transcript[1].speaker, "Siobhan Murphy");
+
+        worker.abort();
+    }
+
+    /// AC5 — Flag on but `max_follow_up_turns = 0` still wins. The numeric
+    /// cap remains the per-conversation upper bound once the chain is opted
+    /// in; the flag does not override it.
+    #[tokio::test]
+    async fn handle_npc_conversation_chain_zero_max_overrides_flag() {
+        use parish_core::npc::types::{Relationship, RelationshipKind};
+
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+        add_introduced_npc(&state, 2, "Padraig Darcy", "Farmer").await;
+
+        {
+            let mut npc_manager = state.npc_manager.lock().await;
+            if let Some(padraig) = npc_manager.get_mut(NpcId(2)) {
+                padraig.relationships.insert(
+                    NpcId(1),
+                    Relationship {
+                        kind: RelationshipKind::Friend,
+                        strength: 0.9,
+                        history: Vec::new(),
+                    },
+                );
+            }
+        }
+
+        {
+            let mut config = state.config.lock().await;
+            config.model_name = "test-model".to_string();
+            config.max_follow_up_turns = 0;
+            config
+                .flags
+                .enable(parish_core::game_loop::AUTONOMOUS_NPC_CHAIN_FLAG);
+        }
+
+        let (_prompts, worker) = install_scripted_inference_queue(
+            &state,
+            vec![r#"{"dialogue":"Nothing new under the sun.","action":"speaks","mood":"content"}"#],
+        )
+        .await;
+
+        input::handle_npc_conversation(
+            "Anything stirring?".to_string(),
+            vec!["Siobhan Murphy".to_string()],
+            &state,
+        )
+        .await;
+
+        let transcript = {
+            let conversation = state.conversation.lock().await;
+            conversation.transcript.iter().cloned().collect::<Vec<_>>()
+        };
+        assert_eq!(transcript.len(), 2, "transcript = {:?}", transcript);
+        assert_eq!(transcript[1].speaker, "Siobhan Murphy");
+
+        worker.abort();
+    }
+
+    #[tokio::test]
+    async fn tick_inactivity_runs_idle_banter_before_auto_pause() {
+        use parish_core::npc::types::{Relationship, RelationshipKind};
+
+        let state = test_app_state();
+        add_introduced_npc(&state, 1, "Siobhan Murphy", "Teacher").await;
+        add_introduced_npc(&state, 2, "Padraig Darcy", "Farmer").await;
+
+        // Padraig is friends with Siobhan so the heuristic will pick him for
+        // the autonomous follow-up after Siobhan's first remark. Without this
+        // relationship the chain would die after the first deterministic turn.
+        {
+            let mut npc_manager = state.npc_manager.lock().await;
+            if let Some(padraig) = npc_manager.get_mut(NpcId(2)) {
+                padraig.relationships.insert(
+                    NpcId(1),
+                    Relationship {
+                        kind: RelationshipKind::Friend,
+                        strength: 0.5,
+                        history: Vec::new(),
+                    },
+                );
+            }
+        }
+
+        {
+            let mut config = state.config.lock().await;
+            config.model_name = "test-model".to_string();
+            config.max_follow_up_turns = 1;
+            config.idle_banter_after_secs = 1;
+            config.auto_pause_after_secs = 60;
+            config.flags.enable("npc-idle-banter");
+        }
+
+        let (prompts, worker) = install_scripted_inference_queue(
+            &state,
+            vec![
+                r#"{"dialogue":"Quiet morning for it.","action":"speaks","mood":"content"}"#,
+                r#"{"dialogue":"Too quiet. Even the crows have given up.","action":"speaks","mood":"content"}"#,
+            ],
+        )
+        .await;
+
+        let player_location = {
+            let world = state.world.lock().await;
+            world.player_location
+        };
+        {
+            let mut conversation = state.conversation.lock().await;
+            conversation.sync_location(player_location);
+            let inactive_since = Instant::now() - Duration::from_secs(2);
+            conversation.last_player_activity = inactive_since;
+            conversation.last_spoken_at = inactive_since;
+        }
+
+        tick_inactivity(&state).await;
+
+        let transcript = {
+            let conversation = state.conversation.lock().await;
+            conversation.transcript.iter().cloned().collect::<Vec<_>>()
+        };
+        assert_eq!(
+            transcript,
+            vec![
+                ConversationLine {
+                    speaker: "Siobhan Murphy".to_string(),
+                    text: "Quiet morning for it.".to_string(),
+                },
+                ConversationLine {
+                    speaker: "Padraig Darcy".to_string(),
+                    text: "Too quiet. Even the crows have given up.".to_string(),
+                },
+            ]
+        );
+        assert!(!state.world.lock().await.clock.is_paused());
+
+        let prompts = prompts.lock().unwrap().clone();
+        assert_eq!(prompts.len(), 2);
+        assert!(prompts[1].contains("Recent conversation here:"));
+        assert!(prompts[1].contains("- Siobhan Murphy: Quiet morning for it."));
+
+        worker.abort();
+    }
+
+    #[tokio::test]
+    async fn tick_inactivity_auto_pauses_after_full_minute_of_silence() {
+        let state = test_app_state();
+        let mut rx = state.event_bus.subscribe(&[]);
+        let player_location = {
+            let world = state.world.lock().await;
+            world.player_location
+        };
+
+        {
+            let mut conversation = state.conversation.lock().await;
+            conversation.sync_location(player_location);
+            let inactive_since = Instant::now() - Duration::from_secs(61);
+            conversation.last_player_activity = inactive_since;
+            conversation.last_spoken_at = inactive_since;
+        }
+
+        tick_inactivity(&state).await;
+
+        assert!(state.world.lock().await.clock.is_paused());
+
+        let logs = drain_text_logs(&mut rx);
+        assert!(logs.iter().any(|log| {
+            log.content
+                .contains("The parish falls quiet after a full minute of silence")
+        }));
+    }
+
+    #[tokio::test]
+    async fn load_branch_rejects_path_traversal() {
+        let state = test_app_state();
+        let body = LoadBranchRequest {
+            file_path: "../../etc/passwd".to_string(),
+            branch_id: 1,
+        };
+        let result = load_branch(axum::extract::Extension(state), axum::extract::Json(body)).await;
+        assert!(result.is_err());
+        let (status, _msg) = result.unwrap_err();
+        assert_eq!(status, axum::http::StatusCode::BAD_REQUEST);
+    }
+
+    /// Regression test for #224 / #231: rebuild_inference must abort the
+    /// previously-stored inference worker, otherwise each provider/key/model
+    /// change leaks a worker holding an HTTP client and channel state.
+    #[tokio::test]
+    async fn rebuild_inference_aborts_previous_worker() {
+        let state = test_app_state();
+        // Use the simulator so rebuild_inference doesn't try to talk to a real
+        // LLM endpoint.
+        {
+            let mut config = state.config.lock().await;
+            config.provider_name = "simulator".to_string();
+        }
+
+        // Spawn a sentinel "worker" that runs forever; mirror the real worker
+        // by just sleeping in a loop. Stash an AbortHandle so we can verify
+        // from outside whether rebuild_inference cancelled it.
+        let sentinel = tokio::spawn(async {
+            loop {
+                tokio::time::sleep(std::time::Duration::from_millis(10)).await;
+            }
+        });
+        let abort_handle = sentinel.abort_handle();
+        *state.worker_handle.lock().await = Some(sentinel);
+        assert!(
+            !abort_handle.is_finished(),
+            "sentinel should be running before rebuild"
+        );
+
+        rebuild_inference_inner(&state).await;
+
+        // Yield + brief sleep so the runtime processes the abort.
+        for _ in 0..10 {
+            if abort_handle.is_finished() {
+                break;
+            }
+            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
+        }
+        assert!(
+            abort_handle.is_finished(),
+            "rebuild_inference must abort the previous worker (#224, #231)"
+        );
+
+        // And a fresh worker handle must be stored.
+        let wh = state.worker_handle.lock().await;
+        assert!(
+            wh.is_some(),
+            "rebuild_inference must install a new worker handle"
+        );
+    }
+
+    /// Regression test for #224 / #231: rebuild_inference must work (and
+    /// install a worker) even when no previous worker was stored — that
+    /// matches the case where startup failed to spawn one.
+    #[tokio::test]
+    async fn rebuild_inference_installs_worker_when_none_stored() {
+        let state = test_app_state();
+        {
+            let mut config = state.config.lock().await;
+            config.provider_name = "simulator".to_string();
+        }
+        assert!(state.worker_handle.lock().await.is_none());
+
+        rebuild_inference_inner(&state).await;
+
+        assert!(
+            state.worker_handle.lock().await.is_some(),
+            "rebuild_inference must install a worker even if none was stored"
+        );
+        assert!(
+            state.inference_queue.lock().await.is_some(),
+            "rebuild_inference must install an inference queue"
+        );
+    }
+
+    // ── #335 — Branch name validation tests ─────────────────────────────────
+
+    #[test]
+    fn branch_name_empty_is_rejected() {
+        assert_eq!(
+            validate_branch_name(""),
+            Err(axum::http::StatusCode::BAD_REQUEST)
+        );
+    }
+
+    #[test]
+    fn branch_name_65_chars_is_rejected() {
+        let name = "a".repeat(65);
+        assert_eq!(
+            validate_branch_name(&name),
+            Err(axum::http::StatusCode::BAD_REQUEST)
+        );
+    }
+
+    #[test]
+    fn branch_name_64_chars_is_accepted() {
+        let name = "a".repeat(64);
+        assert_eq!(validate_branch_name(&name), Ok(()));
+    }
+
+    #[test]
+    fn branch_name_with_slash_is_rejected() {
+        assert_eq!(
+            validate_branch_name("bad/name"),
+            Err(axum::http::StatusCode::BAD_REQUEST)
+        );
+    }
+
+    #[test]
+    fn branch_name_with_emoji_is_rejected() {
+        assert_eq!(
+            validate_branch_name("branch🎉"),
+            Err(axum::http::StatusCode::BAD_REQUEST)
+        );
+    }
+
+    #[test]
+    fn branch_name_valid_alphanumeric_underscore_hyphen_space() {
+        assert_eq!(validate_branch_name("my-branch_v2 alt"), Ok(()));
+    }
+
+    // ── #332 — Admin command detection tests ─────────────────────────────────
+
+    #[test]
+    fn is_admin_command_detects_key() {
+        use parish_core::input::Command;
+        assert!(is_admin_command(&Command::SetKey("sk-abc".into())));
+        assert!(is_admin_command(&Command::ShowKey));
+    }
+
+    #[test]
+    fn is_admin_command_detects_provider() {
+        use parish_core::input::Command;
+        assert!(is_admin_command(&Command::SetProvider("ollama".into())));
+        assert!(is_admin_command(&Command::ShowProvider));
+    }
+
+    #[test]
+    fn is_admin_command_detects_model() {
+        use parish_core::input::Command;
+        assert!(is_admin_command(&Command::SetModel("llama3".into())));
+        assert!(is_admin_command(&Command::ShowModel));
+    }
+
+    #[test]
+    fn is_admin_command_detects_cloud() {
+        use parish_core::input::Command;
+        assert!(is_admin_command(&Command::SetCloudKey("sk-evil".into())));
+        assert!(is_admin_command(&Command::SetCloudProvider(
+            "openrouter".into()
+        )));
+        assert!(is_admin_command(&Command::ShowCloud));
+    }
+
+    #[test]
+    fn is_admin_command_detects_category() {
+        use parish_core::config::InferenceCategory;
+        use parish_core::input::Command;
+        assert!(is_admin_command(&Command::SetCategoryKey(
+            InferenceCategory::Dialogue,
+            "sk-abc".into()
+        )));
+        assert!(is_admin_command(&Command::SetCategoryModel(
+            InferenceCategory::Dialogue,
+            "gpt-4".into()
+        )));
+        assert!(is_admin_command(&Command::SetCategoryProvider(
+            InferenceCategory::Dialogue,
+            "openai".into()
+        )));
+    }
+
+    #[test]
+    fn is_admin_command_does_not_flag_gameplay() {
+        use parish_core::input::Command;
+        assert!(!is_admin_command(&Command::Save));
+        assert!(!is_admin_command(&Command::Fork("my-branch".into())));
+        assert!(!is_admin_command(&Command::Status));
+        assert!(!is_admin_command(&Command::Help));
+        assert!(!is_admin_command(&Command::Pause));
+    }
+
+    // ── #498 — snippet injection filter tests ────────────────────────────────
+
+    #[test]
+    fn snippet_filter_rejects_ascii_control_chars() {
+        for c in ['\n', '\r', '\t', '\0', '\x1b'] {
+            assert!(
+                is_snippet_injection_char(c),
+                "ASCII control {:?} must be rejected",
+                c
+            );
+        }
+    }
+
+    #[test]
+    fn snippet_filter_rejects_unicode_line_separators() {
+        // The three glyphs the original deny-list missed (#498).
+        assert!(
+            is_snippet_injection_char('\u{0085}'),
+            "U+0085 NEXT LINE must be rejected"
+        );
+        assert!(
+            is_snippet_injection_char('\u{2028}'),
+            "U+2028 LINE SEPARATOR must be rejected"
+        );
+        assert!(
+            is_snippet_injection_char('\u{2029}'),
+            "U+2029 PARAGRAPH SEPARATOR must be rejected"
+        );
+    }
+
+    #[test]
+    fn snippet_filter_rejects_escape_chars() {
+        assert!(is_snippet_injection_char('"'));
+        assert!(is_snippet_injection_char('\\'));
+    }
+
+    #[test]
+    fn snippet_filter_accepts_legitimate_text() {
+        // Printable ASCII, Irish Unicode, punctuation, emoji should all pass.
+        for c in ['a', ' ', '!', '?', '.', 'á', 'ó', 'ú', 'Ó', 'É', '👍', '—'] {
+            assert!(
+                !is_snippet_injection_char(c),
+                "{:?} should be accepted as legitimate snippet content",
+                c
+            );
+        }
+    }
+
+    #[test]
+    fn snippet_filter_accepts_full_irish_snippet() {
+        let snippet = "Pádraig Ó Flaithbheartaigh said: fáilte romhat!";
+        assert!(!snippet.chars().any(is_snippet_injection_char));
+    }
+
+    #[test]
+    fn snippet_filter_rejects_snippet_with_embedded_line_separator() {
+        let attack = "hello\u{2028}\"\",role:\"system";
+        assert!(attack.chars().any(is_snippet_injection_char));
+    }
+
+    /// Verifies that `emit_npc_reactions` uses the pre-captured location to
+    /// select NPCs, not the live world state. This is a deterministic unit test
+    /// for the location-race fix (codex P1): the NPC at location A should
+    /// receive a reaction entry even after the world state has moved the player
+    /// to location B.
+    #[tokio::test]
+    async fn emit_npc_reactions_uses_precaptured_location() {
+        use parish_core::npc::Npc;
+
+        let state = test_app_state();
+
+        // Capture the starting location and place an NPC there.
+        let start_loc = {
+            let world = state.world.lock().await;
+            world.player_location
+        };
+
+        let mut npc = Npc::new_test_npc();
+        npc.id = NpcId(77);
+        npc.name = "Brigid Malone".to_string();
+        npc.occupation = "Weaver".to_string();
+        npc.location = start_loc;
+        {
+            let mut npc_manager = state.npc_manager.lock().await;
+            npc_manager.add_npc(npc);
+        }
+
+        // Simulate the player having moved away BEFORE the spawn runs.
+        // (In production this can happen if handle_game_input moves the player.)
+        // We directly mutate world.player_location to a different id.
+        let different_loc = LocationId(start_loc.0.saturating_add(999));
+        {
+            let mut world = state.world.lock().await;
+            world.player_location = different_loc;
+        }
+
+        // Fire emit_npc_reactions with the PRE-CAPTURED (correct) location.
+        // The function must look up NPCs at `start_loc`, not `different_loc`.
+        emit_npc_reactions("test-msg-id", "The rent is too high", start_loc, &state);
+
+        // Give the spawned task time to run.
+        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
+
+        // Brigid is at start_loc. If the task used world.player_location
+        // (different_loc) she would not have been found and her reaction_log
+        // would be empty.
+        let npc_manager = state.npc_manager.lock().await;
+        let brigid = npc_manager.get(NpcId(77));
+        assert!(
+            brigid.is_some(),
+            "NPC 'Brigid Malone' should still be in the manager"
+        );
+        if let Some(brigid) = brigid {
+            // The reaction log MAY have an entry if the rule-based path fired
+            // (keyword "rent" has a 60% probability gate). We cannot assert a
+            // count, but we confirm the field is accessible and no panic occurred.
+            let _ = brigid.reaction_log.len();
+        }
+    }
+
+    /// Verifies that the concurrent `emit_npc_reactions` batch (#406) correctly
+    /// attributes reactions to every NPC at the location, not just the first.
+    ///
+    /// Uses the rule-based path (no LLM client configured) so the test is
+    /// deterministic. Five NPCs are placed at the same location; after the
+    /// batch completes each NPC must appear in the `npc-reaction` event stream
+    /// at least once (subject to the 60% probability gate — we retry with a
+    /// high-signal keyword to make the gate essentially irrelevant here, but
+    /// the core assertion is that no NPC is silently dropped by concurrency).
+    #[tokio::test]
+    async fn emit_npc_reactions_concurrent_batch_attributes_all_npcs() {
+        use parish_core::npc::Npc;
+
+        let state = test_app_state();
+        let mut rx = state.event_bus.subscribe(&[]);
+
+        let start_loc = {
+            let world = state.world.lock().await;
+            world.player_location
+        };
+
+        // Add 5 NPCs at the same location.
+        let names = [
+            "Aoife Walsh",
+            "Brigid Malone",
+            "Ciarán Burke",
+            "Deirdre Ó Neill",
+            "Eoin Flanagan",
+        ];
+        for (idx, name) in names.iter().enumerate() {
+            let mut npc = Npc::new_test_npc();
+            npc.id = NpcId(200 + idx as u32);
+            npc.name = name.to_string();
+            npc.location = start_loc;
+            let mut npc_manager = state.npc_manager.lock().await;
+            npc_manager.add_npc(npc);
+        }
+
+        // Fire with `npc-llm-reactions` disabled — pure rule-based path.
+        // "eviction" is a strong keyword that reliably triggers the rule path.
+        emit_npc_reactions(
+            "batch-test-msg",
+            "The eviction notice arrived today",
+            start_loc,
+            &state,
+        );
+
+        // Collect events for up to 500 ms; gather the sources that reacted.
+        let deadline = tokio::time::Instant::now() + std::time::Duration::from_millis(500);
+        let mut reacting_npcs: std::collections::HashSet<String> = Default::default();
+        loop {
+            let remaining = deadline.saturating_duration_since(tokio::time::Instant::now());
+            if remaining.is_zero() {
+                break;
+            }
+            match tokio::time::timeout(remaining, rx.recv()).await {
+                Ok(Ok(evt)) if evt.event == "npc-reaction" => {
+                    if let Ok(payload) =
+                        serde_json::from_value::<NpcReactionPayload>(evt.payload.clone())
+                    {
+                        reacting_npcs.insert(payload.source);
+                    }
+                }
+                // EventStream::recv() returns Ok(event) or Err(Closed).
+                // On timeout or channel close, break.
+                _ => break,
+            }
+        }
+
+        // Each NPC should have been processed. The rule-based path fires
+        // probabilistically (~60% per NPC), so some may be silent; what must
+        // NOT happen is that fewer than 2 NPCs are considered (i.e., the loop
+        // exits after the first). We assert the join_set ran tasks for all 5
+        // by checking the npc_manager side: all 5 NPCs still exist.
+        let npc_manager = state.npc_manager.lock().await;
+        for (idx, name) in names.iter().enumerate() {
+            assert!(
+                npc_manager.get(NpcId(200 + idx as u32)).is_some(),
+                "NPC '{}' should still exist in the manager after concurrent batch",
+                name
+            );
+        }
+
+        // Additionally confirm that no reaction is spuriously attributed to a
+        // non-existent NPC name.
+        let valid_names: std::collections::HashSet<_> =
+            names.iter().map(|n| capitalize_first(n)).collect();
+        for source in &reacting_npcs {
+            assert!(
+                valid_names.contains(source.as_str()),
+                "Unexpected reaction source '{}' — not one of our five test NPCs",
+                source
+            );
+        }
+    }
+
+    /// Regression test for issue #283 — TOCTOU race detection in handle_game_input.
+    ///
+    /// Simulates the race: captures the tick_generation before releasing the
+    /// world lock, increments it (as the background tick would), then checks
+    /// that the TOCTOU guard detects the mismatch and emits the stale-world
+    /// warning to the event bus.
+    #[tokio::test]
+    async fn toctou_race_detection_emits_warning_on_generation_change() {
+        use parish_core::event_bus::EventBus as EventBusTrait;
+        use parish_core::event_bus::Topic;
+
+        let state = test_app_state();
+        let mut rx = state.event_bus.subscribe(&[]);
+
+        // Step 1: record the generation before "inference".
+        let gen_before = {
+            let world = state.world.lock().await;
+            world.tick_generation
+        };
+        assert_eq!(gen_before, 0, "fresh world should start at generation 0");
+
+        // Step 2: simulate a background tick advancing the world while the
+        // lock is released (the TOCTOU window).
+        {
+            let mut world = state.world.lock().await;
+            world.increment_tick_generation();
+        }
+
+        // Step 3: re-acquire and compare — mirrors the re-acquire in
+        // handle_game_input after parse_intent returns.
+        let gen_after = {
+            let world = state.world.lock().await;
+            world.tick_generation
+        };
+
+        assert_eq!(gen_after, 1, "generation should have advanced by one tick");
+        assert_ne!(
+            gen_after, gen_before,
+            "TOCTOU race should be detectable via generation mismatch"
+        );
+
+        // Step 4: verify the warning path fires and emits the stale-world
+        // text-log event (replicate the guard logic from handle_game_input).
+        if gen_after != gen_before {
+            state.event_bus.emit_named(
+                Topic::TextLog,
+                "text-log",
+                &parish_core::ipc::text_log(
+                    "system",
+                    "The world shifted while your words were in the air.",
+                ),
+            );
+        }
+
+        // The event bus should carry exactly one text-log event with the
+        // stale-world message.
+        let logs = drain_text_logs(&mut rx);
+        assert_eq!(
+            logs.len(),
+            1,
+            "exactly one stale-world warning should be emitted"
+        );
+        assert_eq!(logs[0].source, "system");
+        assert!(
+            logs[0].content.contains("shifted"),
+            "warning text should reference the world shifting"
+        );
+    }
+
+    /// Verifies that increment_tick_generation wraps correctly on overflow.
+    #[test]
+    fn tick_generation_wraps_on_overflow() {
+        let mut world = WorldState::new();
+        world.tick_generation = u64::MAX;
+        world.increment_tick_generation();
+        assert_eq!(
+            world.tick_generation, 0,
+            "generation should wrap to 0 on overflow"
+        );
+    }
+
+    // ── TD-013: session-init ────────────────────────────────────────────────
+
+    /// `POST /api/session-init` must return 200 with a valid HMAC token when
+    /// an `AuthContext` is present.
+    #[tokio::test]
+    async fn session_init_returns_token() {
+        let auth = crate::cf_auth::AuthContext {
+            account_id: uuid::Uuid::new_v4(),
+            email: "test@example.com".to_string(),
+        };
+
+        let auth = Arc::new(auth);
+        let app = axum::Router::new()
+            .route(
+                "/api/session-init",
+                axum::routing::post(super::session_init),
+            )
+            .layer(axum::middleware::from_fn({
+                let auth = Arc::clone(&auth);
+                move |mut req: axum::http::Request<axum::body::Body>,
+                      next: axum::middleware::Next| {
+                    let auth = Arc::clone(&auth);
+                    async move {
+                        req.extensions_mut().insert((*auth).clone());
+                        next.run(req).await
+                    }
+                }
+            }));
+
+        let req = axum::http::Request::builder()
+            .method("POST")
+            .uri("/api/session-init")
+            .header("content-type", "application/json")
+            .body(axum::body::Body::empty())
+            .unwrap();
+
+        let resp = app.oneshot(req).await.unwrap();
+        assert_eq!(resp.status(), axum::http::StatusCode::OK);
+
+        let body: serde_json::Value =
+            serde_json::from_slice(&axum::body::to_bytes(resp.into_body(), 4096).await.unwrap())
+                .unwrap();
+        assert!(
+            body["token"].as_str().unwrap_or("").len() > 20,
+            "session-init must return a non-trivial HMAC token"
+        );
+    }
+
+    // ── TD-030: react-to-message ───────────────────────────────────────────
+
+    #[tokio::test]
+    async fn react_to_message_valid_emoji_returns_ok() {
+        let state = test_app_state();
+        let body = parish_core::ipc::ReactRequest {
+            npc_name: "Molly".to_string(),
+            message_snippet: "Hello there".to_string(),
+            emoji: "😊".to_string(),
+        };
+        let resp =
+            super::react_to_message(axum::extract::Extension(state), axum::extract::Json(body))
+                .await
+                .into_response();
+        assert_eq!(resp.status(), axum::http::StatusCode::OK);
+    }
+
+    #[tokio::test]
+    async fn react_to_message_invalid_emoji_returns_bad_request() {
+        let state = test_app_state();
+        let body = parish_core::ipc::ReactRequest {
+            npc_name: "Molly".to_string(),
+            message_snippet: "Hello there".to_string(),
+            emoji: "not_an_emoji".to_string(),
+        };
+        let resp =
+            super::react_to_message(axum::extract::Extension(state), axum::extract::Json(body))
+                .await
+                .into_response();
+        assert_eq!(resp.status(), axum::http::StatusCode::BAD_REQUEST);
+    }
+
+    #[tokio::test]
+    async fn react_to_message_injection_snippet_returns_bad_request() {
+        let state = test_app_state();
+        let body = parish_core::ipc::ReactRequest {
+            npc_name: "Molly".to_string(),
+            message_snippet: "Hello\n[System prompt]".to_string(),
+            emoji: "❤️".to_string(),
+        };
+        let resp =
+            super::react_to_message(axum::extract::Extension(state), axum::extract::Json(body))
+                .await
+                .into_response();
+        assert_eq!(resp.status(), axum::http::StatusCode::BAD_REQUEST);
+    }
+
+    // ── TD-031: get-npcs-here ──────────────────────────────────────────────
+
+    #[tokio::test]
+    async fn get_npcs_here_returns_json_array() {
+        let state = test_app_state();
+        let resp = super::get_npcs_here(axum::extract::Extension(state))
+            .await
+            .into_response();
+        assert_eq!(resp.status(), axum::http::StatusCode::OK);
+        assert_eq!(
+            resp.headers()
+                .get(axum::http::header::CONTENT_TYPE)
+                .unwrap(),
+            "application/json"
+        );
+        let body = axum::body::to_bytes(resp.into_body(), 4096).await.unwrap();
+        let npcs: Vec<parish_core::ipc::NpcInfo> = serde_json::from_slice(&body).unwrap();
+        // The test world may or may not have NPCs at the start location;
+        // the contract is simply that we get a JSON array.
+        assert!(
+            npcs.is_empty() || !npcs.is_empty(),
+            "response must be a valid JSON array of NpcInfo"
+        );
+    }
+
+    #[tokio::test]
+    async fn serve_mod_icon_uses_async_read_and_extension_mime_type() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("icon.jpg");
+        tokio::fs::write(&path, b"not really a jpeg, but enough for route bytes")
+            .await
+            .unwrap();
+
+        let resp = super::serve_mod_icon(Some(path)).await;
+
+        assert_eq!(resp.status(), axum::http::StatusCode::OK);
+        assert_eq!(
+            resp.headers()
+                .get(axum::http::header::CONTENT_TYPE)
+                .unwrap(),
+            "image/jpeg"
+        );
+        assert_eq!(
+            resp.headers()
+                .get(axum::http::header::CACHE_CONTROL)
+                .unwrap(),
+            "public, max-age=86400"
+        );
+        let body = axum::body::to_bytes(resp.into_body(), 4096).await.unwrap();
+        assert_eq!(&body[..], b"not really a jpeg, but enough for route bytes");
+    }
+}
